requirements.txt
PyYAML==6.0.2
fastapi==0.115.13
uvicorn[standard]==0.32.1
httpx==0.28.1
beautifulsoup4==4.12.3
